+++ ExtremeBowling/src/menu/instructionsMenu.h
@@ -7,7 +7,7 @@ class InstructionsMenu : public Menu {
     public:
         InstructionsMenu(int windowX, int windowY);
         void display();
-        bool backClicked(int x, int y);
+        bool backClicked(float x, float y);
     private: 
         Button back;
 };
